Istall script refactoring to new school bash style
The refactored code does the same but in a new style bash, in less lines of code and a more clean structure. It also doesn't detect non `install.sh` related things like `GCP_PROJECT_NAME`. It is `gcp` package's responsible. The only difference is this version doesn't log `Cloud provider name is missing. Will deploy it on GCP by default.`